# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Financiamento Veículos",
        "description": "API para gerenciamento do sistema de financiamento de veículos",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/financiamentos": {
            "get": {
                "tags": [
                    "Financiamentos"
                ],
                "summary": "Lista todos os financiamentos",
                "operationId": "listarFinanciamentos",
                "responses": {
                    "200": {
                        "description": "Lista retornada com sucesso"
                    }
                }
            },
            "post": {
                "tags": [
                    "Financiamentos"
                ],
                "summary": "Cadastrar financiamento",
                "operationId": "criarFinanciamento",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "cliente_id",
                                    "veiculo_id",
                                    "numero_contrato",
                                    "valor_veiculo",
                                    "valor_entrada",
                                    "valor_financiado",
                                    "taxa_juros",
                                    "quantidade_parcelas",
                                    "data_contratacao",
                                    "situacao"
                                ],
                                "properties": {
                                    "cliente_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "veiculo_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "numero_contrato": {
                                        "type": "string",
                                        "example": "CTR-2026-0001"
                                    },
                                    "valor_veiculo": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 80000
                                    },
                                    "valor_entrada": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 20000
                                    },
                                    "valor_financiado": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 60000
                                    },
                                    "taxa_juros": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 1.99
                                    },
                                    "quantidade_parcelas": {
                                        "type": "integer",
                                        "example": 48
                                    },
                                    "data_contratacao": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2026-07-20"
                                    },
                                    "situacao": {
                                        "type": "string",
                                        "example": "ATIVO"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Financiamento criado com sucesso"
                    }
                }
            }
        },
        "/api/financiamentos/{id}": {
            "get": {
                "tags": [
                    "Financiamentos"
                ],
                "summary": "Buscar financiamento por ID",
                "operationId": "buscarFinanciamento",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do financiamento",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Financiamento encontrado"
                    },
                    "404": {
                        "description": "Financiamento não encontrado"
                    }
                }
            }
        },
...
        "/api/parcelas/{id}": {
            "get": {
                "tags": [
                    "Parcelas"
                ],
                "summary": "Buscar parcela por ID",
                "operationId": "buscarParcela",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID da parcela",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Parcela encontrada"
                    },
                    "404": {
                        "description": "Parcela não encontrada"
                    }
                }
            }
        },
        "/api/parcelas/{id}/pagar": {
            "post": {
                "tags": [
                    "Parcelas"
                ],
                "summary": "Realiza o pagamento da parcela",
                "operationId": "pagarParcela",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID da parcela",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Parcela paga com sucesso"
                    },
                    "404": {
                        "description": "Parcela não encontrada"
                    }
                }
            }
        },
...
    },
...
}